Move string-comment-scan utilities from plugin/transforms to core/utils
Relocated text scanning utilities from src/plugin/src/transforms/source-text/
to src/core/src/utils/text-scan.ts for better architectural alignment.

Original location: src/plugin/src/transforms/source-text/string-comment-scan.ts
New location: src/core/src/utils/text-scan.ts

Rationale:
- These are low-level text scanning utilities, not AST transforms
- Used by multiple transforms across the plugin (conditional-assignment-sanitizer,
  enum-handling, missing-argument-separator-sanitizer)
- No dependencies on plugin-specific logic or Prettier APIs
- Better fit in the core utilities module where other text processing helpers reside

Changes:
- Moved source file and tests to core workspace
- Updated all imports to use Core namespace
- Exported StringCommentScanState type at package level
- Removed empty source-text directory from plugin transforms
- All existing tests pass, build succeeds
