+++ src/core/src/utils/index.ts
@@ -15,4 +15,5 @@ export * from "./numeric-options.js";
 export * from "./object.js";
 export * from "./regexp.js";
 export * from "./string.js";
+export * from "./text-scan.js";
 export * from "./time.js";
