+++ src/core/src/utils/text-scan.ts
