+++ src/plugin/src/transforms/feather/enum-handling.ts
@@ -4,7 +4,6 @@
  */
 import { Core } from "@gml-modules/core";
 
-import { advanceStringCommentScan, createStringCommentScanState } from "../source-text/string-comment-scan.js";
 import { isIntegerLiteralString } from "./utils.js";
 
 function sanitizeEnumBodyInitializerStrings(body: string, bodyStartIndex: number, totalRemoved: number) {
@@ -45,10 +44,10 @@ function sanitizeEnumBodyInitializerStrings(body: string, bodyStartIndex: number
 
 function findNextOpenBrace(sourceText: string, startIndex: number) {
     const length = sourceText.length;
-    const state = createStringCommentScanState();
+    const state = Core.createStringCommentScanState();
 
     for (let index = startIndex; index < length; ) {
-        const nextIndex = advanceStringCommentScan(sourceText, length, index, state, true);
+        const nextIndex = Core.advanceStringCommentScan(sourceText, length, index, state, true);
         if (nextIndex !== index) {
             index = nextIndex;
             continue;
@@ -67,10 +66,10 @@ function findNextOpenBrace(sourceText: string, startIndex: number) {
 function findMatchingClosingBrace(sourceText: string, openBraceIndex: number) {
     const length = sourceText.length;
     let depth = 0;
-    const state = createStringCommentScanState();
+    const state = Core.createStringCommentScanState();
 
     for (let index = openBraceIndex; index < length; ) {
-        const nextIndex = advanceStringCommentScan(sourceText, length, index, state, true);
+        const nextIndex = Core.advanceStringCommentScan(sourceText, length, index, state, true);
         if (nextIndex !== index) {
             index = nextIndex;
             continue;
