--- src/plugin/src/transforms/source-text/index.ts
@@ -1,1 +1,0 @@
-export * from "./string-comment-scan.js";
